feat: remove components with styled-components and styled system
