+++ apps/web/package.json
@@ -30,8 +30,6 @@
     "react-router-dom": "^6.22.3",
     "react-scripts": "5.0.1",
     "sort-by": "^1.2.0",
-    "styled-components": "^6.1.8",
-    "styled-system": "^5.1.5",
     "typescript": "^5.2.2",
     "web-vitals": "^2.1.4"
   },
