+++ apps/web/src/components/Modal/Backdrop.ts
@@ -1,13 +1,13 @@
-import styled from 'styled-components';
+// import styled from 'styled-components';
 
 
-export const Backdrop = styled('div')`
-  position: fixed;
-  z-index: 1040;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  background-color: #000;
-  opacity: 1;
-`;
+// export const Backdrop = styled('div')`
+//   position: fixed;
+//   z-index: 1040;
+//   top: 0;
+//   bottom: 0;
+//   left: 0;
+//   right: 0;
+//   background-color: #000;
+//   opacity: 1;
+// `;
