+++ apps/web/src/components/layout/Header.tsx
@@ -1,94 +1,106 @@
-import { NavLink, useLocation } from 'react-router-dom';
-import styled, { ThemeProvider } from 'styled-components';
-import { Row } from '@lira-dao/ui';
-import { GlobalStyles, Typography } from '../ui';
-import logo from '../../img/logo-dao.png';
-import button from '../../img/enter-app-3.png';
-import theme from '../../theme';
 import React from 'react';
+import { NavLink, useLocation } from 'react-router-dom';
+import { Box, CssBaseline, Grid, Link, Typography, useTheme } from '@mui/material';
+import { ThemeProvider } from '@mui/material/styles';
+import theme from '../../theme_mui';
+import logo from '../../img/logo-dao.png';
+import button from '../../img/enter-app.png';
 import twitter from '../../img/twitter.svg';
 import discord from '../../img/discord.svg';
 
+interface MenuItemProps {
+  children: React.ReactNode;
+  mx: number;
+  isActive?: boolean ;
+}
 
-const StyledMenuItem = styled(Typography)`
-  color: ${props => props.theme.colors['white-80']};
-  cursor: pointer;
-  font-weight: 400;
-  line-height: 27px;
+const MenuItem: React.FC<MenuItemProps> = ({ children, isActive }) => {
+  const theme = useTheme();
 
-  ::selection {
-    color: inherit;
-    background: transparent;
-    text-shadow: none;
-  }
+  return (
+    <Typography
+      sx={{
+        color: isActive ? theme.palette.primary.main : theme.palette.text.secondary,
+        cursor: 'pointer',
+        fontWeight: 400,
+        fontSize: '18px',
+        lineHeight: '27px',
+        mx: 4,
+        '&:hover': {
+          color: theme.palette.common.white
+        },
+        '&:active': {
+          color: theme.palette.primary.main
+        },
+        '::selection': {
+          color: 'inherit',
+          backgroundColor: 'transparent'
+        },
+        '::moz-selection': {
+          color: 'inherit',
+          backgroundColor: 'transparent'
+        }
+      }}
+    >
+      { children }
+    </Typography>
+  );
+};
 
-  /* For Mozilla Firefox */
-
-  ::-moz-selection {
-    color: inherit;
-    background: transparent;
-    text-shadow: none;
-  }
-
-  &:hover {
-    color: ${props => props.theme.colors.white};
-  }
-
-  &:active {
-    color: ${props => props.theme.colors.primary};
-  }
-`;
-
-
-export function Header() {
+export const Header: React.FC = () => {
   const location = useLocation();
+
+  const isContractsActive = location.pathname === '/contracts';
+  const isDaoTeamActive = location.pathname === '/dao_team';
 
   return (
     <ThemeProvider theme={theme}>
-      <GlobalStyles />
-      <Row
-        height={176}
-        margin={['20px 20px 0', '0 20px', '0 20px', '0 20px']}
-        flexDirection={['column', 'row', 'row', 'row']}
-        justifyContent="space-between"
-        alignItems="center"
-      >
-        <Row width={[1, 1 / 3]} justifyContent={['center', 'flex-start']}>
-          <Row mr={4}>
-            <NavLink to="/">
-              <img src={logo} height={66} alt="lira dao logo" />
-            </NavLink>
-          </Row>
-          <Row>
-            <a href="https://twitter.com/LIRA_DAO" target="_blank" rel="noreferrer" style={{ margin: '16px' }}>
-              <img src={twitter} height={28} alt="lira dao twitter" />
-            </a>
-            <a href="https://discord.gg/fDRBajCB9V" target="_blank" rel="noreferrer" style={{ margin: '16px' }}>
-              <img src={discord} height={28} alt="lira dao discord" />
-            </a>
-          </Row>
-        </Row>
+      <CssBaseline />
 
-        <Row justifyContent="center">
+      <Box sx={{
+        height: 176,
+        m: { xs: '20px 20px 0', sm: '0 20px' },
+        display: 'flex',
+        flexDirection: { xs: 'column', sm: 'column', md: 'row' },
+        justifyContent: 'space-between',
+        alignItems: 'center'
+      }}>
+        <Box sx={{ width: [ 1, 1, 1 / 3, 1 / 3 ], display: 'flex', alignItems: 'center', justifyContent: ['center', 'center', 'center', 'flex-start'], mt: 2 }}>
+          <NavLink to="/">
+            <img src={logo} height={66} alt="lira dao logo" />
+          </NavLink>
+          <Link href="https://twitter.com/LIRA_DAO" target="_blank" rel="noreferrer" sx={{ marginLeft: 6, marginRight: 4 }}>
+            <img src={twitter} height={28} alt="Twitter" />
+          </Link>
+          <Link href="https://discord.gg/fDRBajCB9V" target="_blank" rel="noreferrer">
+            <img src={discord} height={28} alt="Discord" />
+          </Link>
+        </Box>
+        <Box sx={{ width: [ 1, 1, 2 / 3, 2 / 3 ], display: 'flex', justifyContent: 'center', flexWrap: 'nowrap' }}>
           <a href="https://whitepaper.liradao.org" target="_blank" rel="noreferrer">
-            <StyledMenuItem marginX={28}>
+            <MenuItem mx={3}>
               WHITEPAPER
-            </StyledMenuItem>
+            </MenuItem>
           </a>
 
           <NavLink to="/contracts">
-            <StyledMenuItem marginX={28}>
+            <MenuItem mx={3} isActive={isContractsActive}>
               CONTRACTS
-            </StyledMenuItem>
+            </MenuItem>
           </NavLink>
-        </Row>
 
-        <Row width={1 / 3} justifyContent="flex-end" display={['none', 'none', 'none', 'flex']}>
+          <NavLink to="/dao_team">
+            <MenuItem mx={3} isActive={isDaoTeamActive}>
+              DAO TEAM
+            </MenuItem>
+          </NavLink>
+        </Box>
+        <Box sx={{ display: ['none', 'none', 'none', 'flex'], justifyContent: 'flex-end' }}>
           <a href="https://dex.liradao.org" target="_blank" rel="noreferrer">
             <img src={button} width={200} alt="Buy LDT" />
           </a>
-        </Row>
-      </Row>
+        </Box>
+      </Box>
     </ThemeProvider>
   );
-}
+};
