--- apps/web/src/components/ui/ColorWrap.ts
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-import { Colors } from '../../theme';
-
-
-export const ColorWrap = styled.span<{ color: keyof Colors }>`
-  margin: 0;
-  color: ${({ color, theme }) => theme.colors[color]};
-`;
