+++ apps/web/src/components/ui/GlobalStyles.ts
@@ -1,75 +1,75 @@
-import { createGlobalStyle } from 'styled-components';
-import ApercuMonoProBold from '../../fonts/ApercuMonoProBold.ttf';
-import ApercuMonoProMedium from '../../fonts/ApercuMonoProMedium.ttf';
-import ApercuMonoProRegular from '../../fonts/ApercuMonoProRegular.ttf';
-import ApercuMonoProLight from '../../fonts/ApercuMonoProLight.ttf';
-import AvenirNextBold from '../../fonts/AvenirNextLTPro-Bold.otf';
-import AvenirNextRegular from '../../fonts/AvenirNextLTPro-Regular.otf';
-import AvenirNextLite from '../../fonts/AvenirNextLTPro-It.otf';
+// import { createGlobalStyle } from 'styled-components';
+// import ApercuMonoProBold from '../../fonts/ApercuMonoProBold.ttf';
+// import ApercuMonoProMedium from '../../fonts/ApercuMonoProMedium.ttf';
+// import ApercuMonoProRegular from '../../fonts/ApercuMonoProRegular.ttf';
+// import ApercuMonoProLight from '../../fonts/ApercuMonoProLight.ttf';
+// import AvenirNextBold from '../../fonts/AvenirNextLTPro-Bold.otf';
+// import AvenirNextRegular from '../../fonts/AvenirNextLTPro-Regular.otf';
+// import AvenirNextLite from '../../fonts/AvenirNextLTPro-It.otf';
 
 
-export const GlobalStyles = createGlobalStyle`
-  @font-face {
-    font-family: "Apercu Mono Pro";
-    src: url(${ApercuMonoProBold}) format("truetype");
-    font-weight: 700;
-  }
+// export const GlobalStyles = createGlobalStyle`
+//   @font-face {
+//     font-family: "Apercu Mono Pro";
+//     src: url(${ApercuMonoProBold}) format("truetype");
+//     font-weight: 700;
+//   }
 
-  @font-face {
-    font-family: "Apercu Mono Pro";
-    src: url(${ApercuMonoProMedium}) format("truetype");
-    font-weight: 300;
-  }
+//   @font-face {
+//     font-family: "Apercu Mono Pro";
+//     src: url(${ApercuMonoProMedium}) format("truetype");
+//     font-weight: 300;
+//   }
 
-  @font-face {
-    font-family: "Apercu Mono Pro";
-    src: url(${ApercuMonoProRegular}) format("truetype");
-    font-weight: 500;
-  }
+//   @font-face {
+//     font-family: "Apercu Mono Pro";
+//     src: url(${ApercuMonoProRegular}) format("truetype");
+//     font-weight: 500;
+//   }
 
-  @font-face {
-    font-family: "Apercu Mono Pro";
-    src: url(${ApercuMonoProLight}) format("truetype");
-    font-weight: 400;
-  }
+//   @font-face {
+//     font-family: "Apercu Mono Pro";
+//     src: url(${ApercuMonoProLight}) format("truetype");
+//     font-weight: 400;
+//   }
 
-  @font-face {
-    font-family: "Avenir Next";
-    src: url(${AvenirNextBold}) format("opentype");
-    font-weight: 700;
-  }
+//   @font-face {
+//     font-family: "Avenir Next";
+//     src: url(${AvenirNextBold}) format("opentype");
+//     font-weight: 700;
+//   }
 
-  @font-face {
-    font-family: "Avenir Next";
-    src: url(${AvenirNextRegular}) format("opentype");
-    font-weight: 500;
-  }
+//   @font-face {
+//     font-family: "Avenir Next";
+//     src: url(${AvenirNextRegular}) format("opentype");
+//     font-weight: 500;
+//   }
 
-  @font-face {
-    font-family: "Avenir Next";
-    src: url(${AvenirNextLite}) format("opentype");
-    font-weight: 400;
-  }
+//   @font-face {
+//     font-family: "Avenir Next";
+//     src: url(${AvenirNextLite}) format("opentype");
+//     font-weight: 400;
+//   }
 
-  html, body {
-    font-family: ${props => props.theme.fontFamilies.primary};
-  }
+//   html, body {
+//     font-family: ${props => props.theme.fontFamilies.primary};
+//   }
 
-  body {
-    min-height: 100vh;
-    margin: 0;
-    background-color: ${props => props.theme.colors.dark};
-    font-family: ${props => props.theme.fontFamilies.primary};
-    -webkit-font-smoothing: antialiased;
-    -moz-osx-font-smoothing: grayscale;
-  }
+//   body {
+//     min-height: 100vh;
+//     margin: 0;
+//     background-color: ${props => props.theme.colors.dark};
+//     font-family: ${props => props.theme.fontFamilies.primary};
+//     -webkit-font-smoothing: antialiased;
+//     -moz-osx-font-smoothing: grayscale;
+//   }
 
-  a,
-  a:visited {
-    text-decoration: none;
-  }
+//   a,
+//   a:visited {
+//     text-decoration: none;
+//   }
 
-  h5 {
-    margin: 0;
-  }
-`;
+//   h5 {
+//     margin: 0;
+//   }
+// `;
