+++ apps/web/src/components/ui/index.ts
@@ -1,3 +1,3 @@
-export * from './GlobalStyles';
-export * from './Typography';
+// export * from './GlobalStyles';
+// export * from './Typography';
 export * from './ColorWrap';
