+++ apps/web/src/pages/Home/components/LiraInfo.tsx
@@ -1,78 +1,56 @@
 import React from 'react';
-import { Col, Row } from '@lira-dao/ui';
-import { ColorWrap, Typography } from '../../../components/ui';
+import { Box, Grid, Typography, Container, ThemeProvider } from '@mui/material';
+import muiTheme from '../../../theme_mui';
+import ColorWrap from '../../../components/ui/ColorWrap';
 import gradientLiraChain from '../../../img/gradient-lira-chain.svg';
 import gradientLiraDao from '../../../img/gradient-lira-dao.svg';
 import gradientIntrinsicValue from '../../../img/gradient-intrinsic-value.svg';
 import gradientLiraDex from '../../../img/gradient-lira-dex.svg';
 
 
-function Image({ img }: { img: string }) {
-  return (
-    <Col
-      alignItems="center"
-      justifyContent="center"
-      marginBottom={[40, 40, 40, 0]}
-      marginX="auto"
-      width={[1, 1 / 2, 1 / 2, 1 / 4]}
-    >
-      <img src={img} width="100%" alt="lira chain" />
-    </Col>
-  );
-}
+const Image = ({ img, alt }: { img: string; alt: string }) => (
+  <Grid xs={10} sm={9} md={7} lg={6} xl={3} sx={{ display: 'flex', alignItems: 'center', justifyContent: 'center' }}>
+    <img src={img} alt={alt} style={{ width: '100%', maxWidth: '80%' }} />
+  </Grid>
+);
 
-function Content({ title, text }: { title: React.ReactNode, text: string }) {
-  return (
-    <Col width={[1, 1, 1, 1 / 2]}>
-      {title}
-
-      <Typography fontSize="xl" textAlign="justify" margin={0}>
+const Content = ({ title, text }: { title: React.ReactNode; text: string }) => (
+  <Grid item xs={12} md={12} lg={6} xl={6}>
+    {title}
+    <Box sx={{ lineHeight: '1.6rem', textAlign: 'justify', width: [1, 1, 1, '92%'] }}>
+      <Typography variant="xl" >
         {text}
       </Typography>
-    </Col>
-  );
-}
+    </Box>
+  </Grid>
+);
 
-const maxWidth = 1280;
-const containerMargin = ['0 40px', '0 40px', '0 40px', '0 4%'];
-const contentMarginBottom = [120, 120, 120, 120];
-
-
-export function LiraInfo() {
+export const LiraInfo: React.FC = () => {
   return (
-    <Col maxWidth={maxWidth} margin="100px auto 0">
-      <Col margin={containerMargin}>
-        <Row flexDirection={['column', 'column', 'column', 'row']} marginBottom={contentMarginBottom}>
-          <Image img={gradientLiraChain} />
-
+    <ThemeProvider theme={muiTheme}>
+      <Container maxWidth="xl" sx={{ 
+        alignItems: 'center',
+        marginTop: '60px',
+        // padding: ['80px 40px 120px', '40px 40px 120px', '60px 40px 120px', '80px auto 120px'],
+        // alignItems: ['center', 'center', 'center', 'center']
+      }}>
+        <Grid container spacing={4} sx={{ mb: '120px', alignItems: 'center', justifyContent: 'center', paddingX: 2 }} >
+          <Image img={gradientLiraChain} alt="Lira Chain" />
           <Content
-            title={
-              <Typography as="h6" margin="0 0 30px">
-                Why <ColorWrap color="primary">Choose</ColorWrap> LIRA CHAIN?
-              </Typography>
-            }
-            text="LIRA CHAIN is a cutting-edge Layer 1 blockchain solution designed to revolutionize decentralized
-              finance (DeFi) and governance. Built with scalability, security, and decentralization at its core, LIRA CHAIN
-              offers a robust infrastructure for executing smart contracts, deploying decentralized applications
-              (DApps), and facilitating secure and efficient transactions. Powered by the native cryptocurrency LIRA DAO
-              COIN (LDC), LIRA CHAIN enables users to participate in governance processes, stake their tokens, and
-              engage in decentralized decision-making within the ecosystem. With its innovative features and
-              community-driven ethos, LIRA CHAIN paves the way for a decentralized future, empowering individuals to
-              take control of their financial destinies and shape the evolution of decentralized ecosystems. Join us
-              on the journey towards a new era of decentralized innovation with LIRA CHAIN."
+            title={<Typography variant="h6" sx={{ fontWeight: 'bold', padding: "0 0 30px" }}>Why <ColorWrap>Choose</ColorWrap> LIRA DEX?</Typography>}
+            text="LIRA DEX stands out in the crowded field of decentralized exchanges by offering a unique and sustainable
+              approach to liquidity mining and reward distribution. Unlike traditional DEXs that calculate rewards based
+              on fiat values, LIRA DEX uses the quantity of LDT tokens to determine rewards, ensuring stability and predictability.
+              Our innovative reward system prevents hyperinflation and hyperdeflation, fostering a healthier market environment. 
+              Additionally, with a significant portion of fees redistributed directly to liquidity providers and a commitment to 
+              sustainable growth, LIRA DEX provides a secure, transparent, and rewarding platform for all participants. 
+              Choose LIRA DEX for a more stable and reliable trading experience."
           />
-        </Row>
-
-        <Row
-          flexDirection={['column-reverse', 'column-reverse', 'column-reverse', 'row']}
-          marginBottom={contentMarginBottom}
-        >
+        </Grid>
+        <Grid container spacing={4} sx={{ mb: '120px', alignItems: 'center', justifyContent: 'center', flexDirection: 'row-reverse' }}>
+          <Image img={gradientLiraDao} alt="Lira Dao" />
           <Content
-            title={
-              <Typography as="h6" margin="0 0 30px">
-                LIRA DAO <ColorWrap color="primary">Ecosystem</ColorWrap>
-              </Typography>
-            }
+            title={<Typography variant="h6" sx={{ fontWeight: 'bold', padding: "0 0 30px" }}>LIRA DAO <ColorWrap>Ecosystem</ColorWrap></Typography>}
             text="LIRA DAO is the decentralized autonomous organization (DAO) driving the governance and development of the
               LIRA ecosystem. Governed by its community of token holders, LIRA DAO empowers participants to collectively
               make decisions regarding protocol upgrades, ecosystem enhancements, and resource allocation. Through
@@ -82,21 +60,11 @@ export function LiraInfo() {
               and community-driven governance. Join us in shaping the future of decentralized finance and governance
               with LIRA DAO."
           />
-
-          <Image img={gradientLiraDao} />
-        </Row>
-      </Col>
-
-      <Col maxWidth={maxWidth} margin={containerMargin}>
-        <Row flexDirection={['column', 'column', 'column', 'row']} marginBottom={contentMarginBottom}>
-          <Image img={gradientIntrinsicValue} />
-
+        </Grid>
+        <Grid container spacing={4} sx={{ mb: '120px', alignItems: 'center', justifyContent: 'center' }}>
+          <Image img={gradientIntrinsicValue} alt="Intrinsic BTC Value" />
           <Content
-            title={
-              <Typography as="h6" margin="0 0 30px">
-                The <ColorWrap color="primary">Power</ColorWrap> of Intrinsic BTC Value
-              </Typography>
-            }
+            title={<Typography variant="h6" sx={{ fontWeight: 'bold', padding: "0 0 30px" }} gutterBottom>The <ColorWrap>Power</ColorWrap> of Intrinsic BTC Value</Typography>}
             text="Satoshi LIRA Token is a groundbreaking digital asset native to the LIRA blockchain, uniquely designed to
               encapsulate the principles of sound money and financial sovereignty. With a fixed supply and intrinsic
               value linked to Bitcoin, Satoshi LIRA Token represents a secure store of value within the LIRA ecosystem.
@@ -105,29 +73,18 @@ export function LiraInfo() {
               inclusivity. Join us in harnessing the power of Satoshi LIRA Token to unlock new possibilities in
               decentralized finance and reshape the future of digital economies."
           />
-        </Row>
-
-        <Row
-          flexDirection={['column-reverse', 'column-reverse', 'column-reverse', 'row']}
-          marginBottom={contentMarginBottom}
-        >
+        </Grid>
+        <Grid container spacing={4} sx={{ mb: '120px', alignItems: 'center', justifyContent: 'center', flexDirection: 'row-reverse' }}>
+          <Image img={gradientLiraDex} alt="Lira DEX" />
           <Content
-            title={
-              <Typography as="h6" margin="0 0 30px">
-                LIRA <ColorWrap color="primary">DEX</ColorWrap>
-              </Typography>
-            }
-            text="LIRA DEX is the decentralized exchange (DEX) built on the LIRA blockchain, providing users with a secure
-              and efficient platform for trading digital assets. Powered by cutting-edge technology and designed with
-              user experience in mind, LIRA DEX offers seamless access to a wide range of cryptocurrencies, facilitating
-              peer-to-peer transactions with minimal fees and maximum security. With its decentralized architecture
-              and non-custodial nature, LIRA DEX empowers users to retain control over their assets while participating in
-              a vibrant and liquid marketplace. Join us on LIRA DEX and experience the future of decentralized trading firsthand."
+            title={<Typography variant="h6" sx={{ fontWeight: 'bold', padding: "0 0 30px" }} gutterBottom>Become an <ColorWrap>Ambassador</ColorWrap></Typography>}
+            text="Join the LIRA DAO Ambassador Program and help expand our vibrant community. As an ambassador, you'll earn rewards
+              through a multi-tiered referral system and contribute to the growth of the LIRA ecosystem. Maintain an active TB
+              token staking position for six months to qualify and start benefiting from our comprehensive incentive structure.
+              Learn more on our whitepaper about how you can get involved and make a difference with LIRA DAO."
           />
-
-          <Image img={gradientLiraDex} />
-        </Row>
-      </Col>
-    </Col>
+        </Grid>
+      </Container>
+    </ThemeProvider>
   );
-}
+};
