+++ apps/web/src/pages/Home/components/PreSaleInfoCard.ts
@@ -1,8 +1,8 @@
-import styled from 'styled-components';
-import { Row } from '@lira-dao/ui';
+// import styled from 'styled-components';
+// import { Row } from '@lira-dao/ui';
 
 
-export const PreSaleInfoCard = styled(Row)<any>`
-  background: url(${props => props.background}) no-repeat left center;
-  background-size: cover;
-`;
+// export const PreSaleInfoCard = styled(Row)<any>`
+//   background: url(${props => props.background}) no-repeat left center;
+//   background-size: cover;
+// `;
