+++ apps/web/src/router.tsx
@@ -1,7 +1,6 @@
 import { createBrowserRouter, Navigate, redirect } from 'react-router-dom';
 import { MainLayout } from './components/layout/MainLayout';
-// import { Home } from './pages/Home';
-import { NewHome } from './pages/Home/NewHome';
+import { Home } from './pages/Home';
 import { Contracts } from './pages/Contracts';
 import { DaoTeam } from './pages/DaoTeam/DaoTeam';
 
@@ -15,7 +14,7 @@ export const router = createBrowserRouter([
     path: '/',
     element: <MainLayout />,
     children: [
-      { index: true, element: <NewHome /> },
+      { index: true, element: <Home /> },
       { path: 'contracts', element: <Contracts /> },
       { path: 'dao_team', element: <DaoTeam /> },
       { path: 'presale', element: <Presale />}
